user model, integerated next-auth v5

+++ src/actions/actions.ts
@@ -3,14 +3,22 @@
 import prisma from "@/lib/db";
 import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { Pet } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
+  const validatedPet = petFormSchema.safeParse(pet);
+
+  if (!validatedPet.success) {
+    return {
+      message: "Invalid Pet data",
+    };
+  }
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (err) {
     return {
@@ -21,14 +29,24 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPet: PetEssentials) {
+export async function editPet(petId: unknown, newPet: unknown) {
   await sleep(1000);
+
+  const validPetId = petIdSchema.safeParse(petId);
+  const validPetData = petFormSchema.safeParse(newPet);
+
+  if (!validPetData.success || !validPetId.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
+
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validPetId.data,
       },
-      data: newPet,
+      data: validPetData.data,
     });
   } catch (error) {
     return {
@@ -39,13 +57,21 @@ export async function editPet(petId: Pet["id"], newPet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validPetId = petIdSchema.safeParse(petId);
+
+  if (!validPetId.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validPetId.data,
       },
     });
   } catch (error) {
